commands/workflows/requirements.txt
# Workflow Framework Requirements

# YAML parsing
PyYAML>=6.0.1

# Graph operations for dependency resolution
networkx>=3.0

# Rich CLI output (optional, for enhanced display)
rich>=13.0.0

# Async operations
asyncio>=3.4.3

# Type hints
typing-extensions>=4.5.0